Transform the repository into standard maven format and merge the pom.xml of the release repo
