+++ src/main/java/org/json/JSONException.java
