+++ src/main/java/org/json/JSONPointer.java
